........
  r1112414 | clicea | 2010-04-08 07:30:16 +0200 (Thu, 08 Apr 2010) | 3 lines
  
  Better magic number to convert from column's units (in 256's of the width of the biggest charecter 
  between 1,2,3...,9) to pts
........
  r1112430 | clicea | 2010-04-08 07:58:16 +0200 (Thu, 08 Apr 2010) | 5 lines
  
  ->Stop torturing myself and store the dimensions in float.
  ->Use the recent improvements in the pt conversion.
  ->Much better positioning. Still not perfect, but it's much better already.
........
  r1112431 | clicea | 2010-04-08 08:04:53 +0200 (Thu, 08 Apr 2010) | 2 lines
  
  ->Wisth is perfect now, height needs further investigation.
........

svn path=/trunk/koffice/; revision=1119719

+++ filters/kspread/xlsx/ChartExport.cpp
@@ -73,10 +73,10 @@ bool ChartExport::saveIndex(KoXmlWriter* xmlWriter)
     //xmlWriter->addAttribute("table:end-x", "0.2953in");
     //xmlWriter->addAttribute("table:end-y", "0.0232in");
     xmlWriter->addAttribute("table:end-cell-address", m_endCellAddress);
-    xmlWriter->addAttributePt("svg:x", (float)m_x);
-    xmlWriter->addAttributePt("svg:y", (float)m_y);
-    xmlWriter->addAttributePt("svg:width", (float)m_width);
-    xmlWriter->addAttributePt("svg:height", (float)m_height);
+    xmlWriter->addAttributePt("svg:x", m_x);
+    xmlWriter->addAttributePt("svg:y", m_y);
+    xmlWriter->addAttributePt("svg:width", m_width);
+    xmlWriter->addAttributePt("svg:height", m_height);
     xmlWriter->addAttribute("draw:z-index", "0");
     xmlWriter->startElement("draw:object");
     xmlWriter->addAttribute("draw:notify-on-update-of-ranges", m_notifyOnUpdateOfRanges);
@@ -111,8 +111,8 @@ bool ChartExport::saveContent(KoStore* store, KoXmlWriter* manifestWriter)
     if (!chart()->m_impl->name().isEmpty())
         bodyWriter->addAttribute("chart:class", "chart:" + chart()->m_impl->name());
 
-    bodyWriter->addAttributePt("svg:width", (float)m_width);
-    bodyWriter->addAttributePt("svg:height", (float)m_height);
+    bodyWriter->addAttributePt("svg:width", m_width);
+    bodyWriter->addAttributePt("svg:height", m_height);
 
     KoGenStyle style(KoGenStyle::GraphicAutoStyle, "chart");
     //style.addProperty("draw:stroke", "none");
